requirements.txt
bcrypt==3.1.7
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
gcloud==0.17.0
googleapis-common-protos==1.6.0
httplib2==0.14.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
jws==0.1.3
MarkupSafe==1.1.1
oauth2client==3.0.0
protobuf==3.10.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycparser==2.19
pycryptodome==3.4.3
PyJWT==1.7.1
Pyrebase==3.0.27
python-jwt==2.0.1
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.0
six==1.13.0
SQLAlchemy==1.3.10
urllib3==1.25.6
Werkzeug==0.16.0
WTForms==2.2.1
